extract course finder to package
