+++ app/src/main/java/org/stepic/droid/core/StepicCoreComponent.java
@@ -16,7 +16,7 @@ import org.stepic.droid.model.Section;
 import org.stepic.droid.notifications.HackFcmListener;
 import org.stepic.droid.notifications.HackerFcmInstanceId;
 import org.stepic.droid.notifications.NotificationBroadcastReceiver;
-import org.stepic.droid.presenters.CourseFinderPresenterImpl;
+import org.stepic.droid.presenters.course_finder.CourseFinderPresenterImpl;
 import org.stepic.droid.receivers.DownloadClickReceiver;
 import org.stepic.droid.receivers.DownloadCompleteReceiver;
 import org.stepic.droid.receivers.InternetConnectionEnabledReceiver;
