+++ app/src/main/java/org/stepic/droid/core/StepicDefaultModule.java
@@ -26,9 +26,9 @@ import org.stepic.droid.notifications.NotificationManagerImpl;
 import org.stepic.droid.notifications.model.Notification;
 import org.stepic.droid.preferences.SharedPreferenceHelper;
 import org.stepic.droid.preferences.UserPreferences;
-import org.stepic.droid.presenters.CourseFinderPresenter;
-import org.stepic.droid.presenters.CourseFinderPresenterForDetailScreen;
-import org.stepic.droid.presenters.CourseFinderPresenterForSectionScreen;
+import org.stepic.droid.presenters.course_finder.CourseFinderPresenter;
+import org.stepic.droid.presenters.course_finder.CourseFinderPresenterForDetailScreen;
+import org.stepic.droid.presenters.course_finder.CourseFinderPresenterForSectionScreen;
 import org.stepic.droid.social.SocialManager;
 import org.stepic.droid.store.CleanManager;
 import org.stepic.droid.store.ConcurrentCancelSniffer;
