+++ app/src/main/java/org/stepic/droid/presenters/course_finder/CourseFinderPresenterForSectionScreen.java
@@ -1,4 +1,4 @@
-package org.stepic.droid.presenters;
+package org.stepic.droid.presenters.course_finder;
 
 import javax.inject.Singleton;
