+++ app/src/main/java/org/stepic/droid/presenters/course_joiner/CourseJoinerPresenter.java
@@ -1,0 +1,4 @@
+package org.stepic.droid.presenters.course_joiner;
+
+public interface CourseJoinerPresenter {
+}
