+++ app/src/main/java/org/stepic/droid/view/abstraction/JoinCourseView.java
@@ -1,0 +1,5 @@
+package org.stepic.droid.view.abstraction;
+
+public interface JoinCourseView {
+    void showProgress();
+}
